requirements.txt
Django==5.1.7
django-axes==6.1.1
django-password-validators==1.7.0
django-csp==3.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==11.1.0
django-debug-toolbar==4.2.0
reportlab==4.0.9
django-filter==23.5
django-imagekit==5.0.0
django-cleanup==8.0.0
python-dotenv>=1.0.0
python-dotenv>=1.0.0
python-sentry-sdk
django-compressor==4.1
whitenoise==6.2.0